+++ Validators/LoginRequestValidator.cs
@@ -1,0 +1,24 @@
+using FluentValidation;
+using BoltWebAPI.Models.Domain;
+
+namespace BoltWebAPI.Validators;
+
+public class LoginRequestValidator : AbstractValidator<LoginRequest>
+{
+    public LoginRequestValidator()
+    {
+        RuleFor(x => x.Username)
+            .NotEmpty()
+            .WithMessage("Username is required")
+            .MinimumLength(3)
+            .WithMessage("Username must be at least 3 characters")
+            .MaximumLength(50)
+            .WithMessage("Username must not exceed 50 characters");
+
+        RuleFor(x => x.Password)
+            .NotEmpty()
+            .WithMessage("Password is required")
+            .MinimumLength(8)
+            .WithMessage("Password must be at least 8 characters");
+    }
+}