+++ Validators/RefreshTokenRequestValidator.cs
@@ -1,0 +1,14 @@
+using FluentValidation;
+using BoltWebAPI.Models.Domain;
+
+namespace BoltWebAPI.Validators;
+
+public class RefreshTokenRequestValidator : AbstractValidator<RefreshTokenRequest>
+{
+    public RefreshTokenRequestValidator()
+    {
+        RuleFor(x => x.RefreshToken)
+            .NotEmpty()
+            .WithMessage("Refresh token is required");
+    }
+}